+++ test/unit/environment/helpers/check_blrp_config.dart
@@ -14,13 +14,19 @@ void main() {
 
   // Convert Duration to milliseconds for JSON serialization
   final jsonConfig = <String, dynamic>{};
-  config.forEach((key, value) {
-    if (value is Duration) {
-      jsonConfig['${key}_ms'] = value.inMilliseconds;
-    } else {
-      jsonConfig[key] = value;
-    }
-  });
+
+  if (config.scheduleDelay != null) {
+    jsonConfig['scheduleDelay_ms'] = config.scheduleDelay!.inMilliseconds;
+  }
+  if (config.exportTimeout != null) {
+    jsonConfig['exportTimeout_ms'] = config.exportTimeout!.inMilliseconds;
+  }
+  if (config.maxQueueSize != null) {
+    jsonConfig['maxQueueSize'] = config.maxQueueSize;
+  }
+  if (config.maxExportBatchSize != null) {
+    jsonConfig['maxExportBatchSize'] = config.maxExportBatchSize;
+  }
 
   print(jsonEncode(jsonConfig));
 }